third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/monitor-console-command-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests traceCalls(fn) console command.


Running: testSimpleMonitor
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction called
Stop monitoring.

Running: testSimpleMonitorWith1Arg
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction called with arguments: 1
Stop monitoring.

Running: testSimpleMonitorWithManyArgs
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction called with arguments: 1, 2, 3, 4, 5
Stop monitoring.

Running: testSimpleUnmonitor
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction3 called
Stop monitoring.

Running: testUnmonitorFuntionNotMonitored
